+++ Configuration/PortalNodeConfigurationInstructionProcessor.php
@@ -8,10 +8,11 @@ use Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\ClosureInstructio
 use Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\Contract\InstructionLoaderInterface;
 use Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\Contract\InstructionTokenContract;
 use Heptacom\HeptaConnect\Core\Configuration\Contract\PortalNodeConfigurationProcessorInterface;
+use Heptacom\HeptaConnect\Core\Portal\Contract\PackageQueryMatcherInterface;
 use Heptacom\HeptaConnect\Core\Portal\Contract\PortalRegistryInterface;
+use Heptacom\HeptaConnect\Portal\Base\Portal\PortalCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
-use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
-use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\PortalNodeKeyCollection;
 use Psr\Log\LoggerInterface;
 
 final class PortalNodeConfigurationInstructionProcessor implements PortalNodeConfigurationProcessorInterface
@@ -20,9 +21,9 @@ final class PortalNodeConfigurationInstructionProcessor implements PortalNodeCon
 
     private LoggerInterface $logger;
 
-    private StorageKeyGeneratorContract $storageKeyGenerator;
-
     private PortalRegistryInterface $portalRegistry;
+
+    private PackageQueryMatcherInterface $packageQueryMatcher;
 
     /**
      * @var InstructionLoaderInterface[]
@@ -34,13 +35,13 @@ final class PortalNodeConfigurationInstructionProcessor implements PortalNodeCon
      */
     public function __construct(
         LoggerInterface $logger,
-        StorageKeyGeneratorContract $storageKeyGenerator,
         PortalRegistryInterface $portalRegistry,
+        PackageQueryMatcherInterface $packageQueryMatcher,
         iterable $instructionLoaders
     ) {
         $this->logger = $logger;
-        $this->storageKeyGenerator = $storageKeyGenerator;
         $this->portalRegistry = $portalRegistry;
+        $this->packageQueryMatcher = $packageQueryMatcher;
         $this->instructionLoaders = \iterable_to_array($instructionLoaders);
     }
 
@@ -71,66 +72,39 @@ final class PortalNodeConfigurationInstructionProcessor implements PortalNodeCon
     private function filterInstructions(PortalNodeKeyInterface $portalNodeKey): array
     {
         $result = [];
+        $portalExtensions = null;
 
         foreach ($this->getInstructions() as $instruction) {
             $query = $instruction->getQuery();
+            $matchedKeys = $this->packageQueryMatcher->matchPortalNodeKeys($query, new PortalNodeKeyCollection([
+                $portalNodeKey,
+            ]));
 
-            if (\class_exists($query) || \interface_exists($query)) {
-                if (
-                    !$this->isQueryMatchingPortalNode($portalNodeKey, $query)
-                    && !$this->isQueryMatchingPortalExtension($portalNodeKey, $query)
-                ) {
-                    continue;
-                }
-            } elseif (!$this->isQueryMatchingPortalNodeKey($portalNodeKey, $query)) {
+            if ($matchedKeys->count() > 0) {
+                $result[] = $instruction;
+
                 continue;
             }
 
-            $result[] = $instruction;
+            $portalExtensions ??= $this->portalRegistry->getPortalExtensions($portalNodeKey);
+            $matchedPortals = $this->packageQueryMatcher->matchPortals($query, new PortalCollection([
+                $this->portalRegistry->getPortal($portalNodeKey),
+            ]));
+
+            if ($matchedPortals->count() > 0) {
+                $result[] = $instruction;
+
+                continue;
+            }
+
+            $matchedPortalExtensions = $this->packageQueryMatcher->matchPortalExtensions($query, $portalExtensions);
+
+            if ($matchedPortalExtensions->count() > 0) {
+                $result[] = $instruction;
+            }
         }
 
         return $result;
-    }
-
-    private function isQueryMatchingPortalNodeKey(PortalNodeKeyInterface $portalNodeKey, string $query): bool
-    {
-        try {
-            return $this->storageKeyGenerator->deserialize($query)->equals($portalNodeKey);
-        } catch (UnsupportedStorageKeyException $e) {
-            if ($this->storageKeyGenerator->serialize($portalNodeKey->withoutAlias()) === $query) {
-                return true;
-            }
-
-            if ($this->storageKeyGenerator->serialize($portalNodeKey->withAlias()) === $query) {
-                return true;
-            }
-
-            return false;
-        }
-    }
-
-    /**
-     * @param class-string $classString
-     */
-    private function isQueryMatchingPortalNode(PortalNodeKeyInterface $portalNodeKey, string $classString): bool
-    {
-        $portal = $this->portalRegistry->getPortal($portalNodeKey);
-
-        return $portal instanceof $classString;
-    }
-
-    /**
-     * @param class-string $classString
-     */
-    private function isQueryMatchingPortalExtension(PortalNodeKeyInterface $portalNodeKey, string $classString): bool
-    {
-        foreach ($this->portalRegistry->getPortalExtensions($portalNodeKey) as $portalExtension) {
-            if ($portalExtension instanceof $classString) {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     /**