+++ Ui/Admin/Action/PortalNodeExtensionActivateUi.php
@@ -5,19 +5,17 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Core\Ui\Admin\Action;
 
 use Heptacom\HeptaConnect\Core\Portal\ComposerPortalLoader;
-use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalExtensionContract;
-use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionCollection;
+use Heptacom\HeptaConnect\Core\Portal\Contract\PackageQueryMatcherInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\PortalNodeKeyCollection;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Activate\PortalExtensionActivatePayload;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Get\PortalNodeGetCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionActivateActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionFindActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNode\PortalNodeGetActionInterface;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayloads;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayload;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\PortalNode\PortalNodeExtensionActivateUiActionInterface;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionIsAlreadyActiveOnPortalNodeException;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionMissingException;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionDoesNotSupportPortalNodeException;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\NoMatchForPackageQueryException;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionsAreAlreadyActiveOnPortalNodeException;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalNodeMissingException;
 
 final class PortalNodeExtensionActivateUi implements PortalNodeExtensionActivateUiActionInterface
@@ -28,51 +26,59 @@ final class PortalNodeExtensionActivateUi implements PortalNodeExtensionActivate
 
     private PortalExtensionActivateActionInterface $portalExtensionActivateAction;
 
+    private PackageQueryMatcherInterface $packageQueryMatcher;
+
     private ComposerPortalLoader $portalLoader;
 
     public function __construct(
         PortalNodeGetActionInterface $portalNodeGetAction,
         PortalExtensionFindActionInterface $portalExtensionFindAction,
         PortalExtensionActivateActionInterface $portalExtensionActivateAction,
+        PackageQueryMatcherInterface $packageQueryMatcher,
         ComposerPortalLoader $portalLoader
     ) {
         $this->portalNodeGetAction = $portalNodeGetAction;
         $this->portalExtensionFindAction = $portalExtensionFindAction;
         $this->portalExtensionActivateAction = $portalExtensionActivateAction;
+        $this->packageQueryMatcher = $packageQueryMatcher;
         $this->portalLoader = $portalLoader;
     }
 
-    public function activate(PortalNodeExtensionActivatePayloads $payloads): void
+    public function activate(PortalNodeExtensionActivatePayload $payload): void
     {
-        $portalNodeKey = $payloads->getPortalNodeKey();
+        if ($payload->getPortalExtensionQueries() === []) {
+            return;
+        }
+
+        $portalNodeKey = $payload->getPortalNodeKey();
         $portalNodeGetCriteria = new PortalNodeGetCriteria(new PortalNodeKeyCollection([$portalNodeKey]));
 
         foreach ($this->portalNodeGetAction->get($portalNodeGetCriteria) as $portalNodeGetResult) {
             $portalExtensionActivePayload = new PortalExtensionActivatePayload($portalNodeKey);
-            $portalExtensions = $this->portalLoader->getPortalExtensions();
+            $portalExtensions = $this->portalLoader->getPortalExtensions()->bySupport($portalNodeGetResult->getPortalClass());
             $portalExtensionState = $this->portalExtensionFindAction->find($portalNodeKey);
+            $alreadyActiveExtensions = [];
 
-            foreach ($payloads->getPortalExtensionClasses() as $portalExtensionClass) {
-                if (!\class_exists($portalExtensionClass)) {
-                    throw new PortalExtensionMissingException($portalExtensionClass, 1650142325);
+            foreach ($payload->getPortalExtensionQueries() as $query) {
+                $queriedPortalExtensions = $this->packageQueryMatcher->matchPortalExtensions($query, $portalExtensions);
+
+                if ($queriedPortalExtensions->count() === 0) {
+                    throw new NoMatchForPackageQueryException($query, 1650142326);
                 }
 
-                $portalClass = $portalNodeGetResult->getPortalClass();
-                $portalExtension = $this->getPortalExtensionByType($portalExtensions->bySupport($portalClass), $portalExtensionClass);
+                foreach ($queriedPortalExtensions as $portalExtension) {
+                    $portalExtensionClass = \get_class($portalExtension);
 
-                if ($portalExtension === null) {
-                    throw new PortalExtensionDoesNotSupportPortalNodeException($portalNodeKey, $portalExtensionClass, 1650142326);
+                    if ($portalExtensionState->isActive($portalExtension)) {
+                        $alreadyActiveExtensions[] = $portalExtensionClass;
+                    } else {
+                        $portalExtensionActivePayload->addExtension($portalExtensionClass);
+                    }
                 }
-
-                if ($portalExtensionState->isActive($portalExtension)) {
-                    throw new PortalExtensionIsAlreadyActiveOnPortalNodeException($portalNodeKey, $portalExtensionClass, 1650142327);
-                }
-
-                $portalExtensionActivePayload->addExtension($portalExtensionClass);
             }
 
             if ($portalExtensionActivePayload->getExtensions() === []) {
-                return;
+                throw new PortalExtensionsAreAlreadyActiveOnPortalNodeException($portalNodeKey, $alreadyActiveExtensions, 1650142327);
             }
 
             $this->portalExtensionActivateAction->activate($portalExtensionActivePayload);
@@ -81,21 +87,5 @@ final class PortalNodeExtensionActivateUi implements PortalNodeExtensionActivate
         }
 
         throw new PortalNodeMissingException($portalNodeKey, 1650142328);
-    }
-
-    /**
-     * @param class-string<PortalExtensionContract> $portalExtensionClass
-     */
-    private function getPortalExtensionByType(
-        PortalExtensionCollection $extensions,
-        string $portalExtensionClass
-    ): ?PortalExtensionContract {
-        foreach ($extensions as $portalExtension) {
-            if ($portalExtension instanceof $portalExtensionClass) {
-                return $portalExtension;
-            }
-        }
-
-        return null;
     }
 }